implement ticket cancellation, event filter search & pagination

+++ src/main/java/com/milos/tickethub/service/EventServiceImpl.java
@@ -1,17 +1,24 @@
 package com.milos.tickethub.service;
 
+import com.milos.tickethub.dto.EventResponse;
 import com.milos.tickethub.entity.Event;
 import com.milos.tickethub.exception.EventNotFoundException;
+import com.milos.tickethub.mapper.EventMapper;
 import com.milos.tickethub.repository.EventRepository;
+import com.milos.tickethub.repository.TicketRepository;
 import lombok.AllArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
 @AllArgsConstructor
 public class EventServiceImpl implements EventService{
     private final EventRepository eventRepository;
+    private final TicketRepository ticketRepository;
 
     @Override
     public List<Event> getAllEvents(){
@@ -36,8 +43,20 @@ public class EventServiceImpl implements EventService{
     }
 
     @Override
+    public Page<EventResponse> findWithFilters(String title, String location, LocalDateTime date, Pageable pageable){
+        String titleParam = (title != null && !title.isBlank()) ? "%" + title + "%" : null;
+        Page<Event> eventPage = eventRepository.findWithFilters(titleParam, location, date, pageable);
+        return eventPage.map(EventMapper::toResponse);
+    }
+
+    @Override
     public String deleteEvent(Long id){
         eventRepository.findById(id).orElseThrow(() -> new EventNotFoundException(id));
+        if(ticketRepository.existsByEventId(id)){
+            throw new IllegalStateException(
+                    "Cannot delete event with existing tickets"
+            );
+        }
         eventRepository.deleteById(id);
         return "Event deleted successfully";
     }
